chore: extract some types
